Set correct minzoom for railway features, add railway subclass
Minzoom for railway features depends on railway=*, service=* and
usage=*. Only railway features with specific railway=* values are added
to the vector tiles. This means, railway=construction/disused/… and
railway=fancy_value are ignored.

The OpenMapTiles schema defines a subclass field for railways as well.
This commit adds this field.