tile_data and child classes: baseZoom -> indexZoom
This is a mechanical renaming, I haven't yet taught tilemaker.cpp how to
clamp basezooms that are greater than 14.

Another performance implication: clip_cache maxes out its clipping at
z14, so we'll do some redundant clipping when the map's basezoom is 15+.
It doesn't have to be this way -- I just didn't want to thread the real
basezoom into TileDataSource just for this case. This could be restored
later, although I imagine in practice, the performance benefit of
caching past z14 is fairly modest.