attempt to support boost 1.86
See issue #747.

The C++ powers that be have realized that something is redundant. The
Boost people have removed this redundancy, and the guidance is to use
the Other Way of Doing The Thing.

The only problem is... I don't know what the Other Way is. From
context I infer there's some static method we ought to be able to
invoke, and it's so obvious that someone who writes custom allocators
would just immediately understand what they're referring to. That's not
me, though.

But...do we even care? Can we just... not call the destroy function?

I think destroy only gets called when deallocate is called. I think we only
call deallocate in the case when a user reads multiple PBFs. I think that's
an uncommon case. I think (big grain of salt - I skimmed, and I'm out of my depth here)
that boost will free up the underlying memory if all of the objects in
a given segment are deallocated. So probably, in the multiple PBF case,
we introduce a memory leak by not calling the destroy function.

That's not great, but it's probably fine? If you're processing a bunch
of small PBFs, you'll probably still be able to do so. If you're
processing large PBFs, you'll still be saved by `--store`.

This is, obviously, an incredibly hacky fix for this issue. But I wanted
to send some PRs for other bugs, and I want to get the build to work
first. :)