make the spatial index sparse
We previously had a dense bitset that contained all the z15 tiles.

Now, we split the z15 tiles up, grouped by z6 tile. We only allocate a
dense bitset for the z6 region if there is at least one z15 tile in the
region.

This is slightly slower at runtime due to the extra indirections, but
the memory savings is likely worth it.