D'oh, ban copy/move constructors
Oops: if you ever passed a non-default compression level, we'd
try to create a new compressor. But because we were using the
default constructors, we'd happily create a temporary, copy its
pointer over to the `compressor` field, then destruct the temporary,
freeing the underlying native resource, and the world would end.

Instead, delete the default constructors and require an explicit
`setLevel` call.

Also update tests to verify that the different levels work.